visa/requirements.txt
altgraph==0.17.3
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.6
ezdxf==1.1.3
filelock==3.12.2
fonttools==4.47.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
MouseInfo==0.1.3
numpy==1.25.0
packaging==23.2
pandas==2.2.0
pefile==2023.2.7
Pillow==10.0.0
pipenv==2023.7.4
platformdirs==3.8.0
pyarrow==15.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyMsgBox==1.0.9
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtChart==5.15.6
PyQtChart-Qt5==5.15.2
pyqtgraph==0.13.3
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
PySimpleGUI==4.60.5
pytesseract==0.3.10
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2024.1
PyVISA==1.13.0
PyVISA-py==0.7.0
pywin32-ctypes==0.2.2
QDarkStyle==3.1
qt-material==2.14
QtPy==2.4.1
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
visa==1.0.0
